refactor(cleanup): cleans up a bit, adds some comments, renames components
