+++ GrasshopperAsyncComponent/GrasshopperAsyncComponent.csproj
@@ -64,6 +64,7 @@
     <Compile Include="Base\IAsyncComponentWorker.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SampleImplementations\Sample_PrimeCalculatorAsyncComponent.cs" />
+    <Compile Include="SampleImplementations\Sample_UslessCyclesComponent.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
